requirements.txt
psycopg2-binary
rich
retry
PyGithub
python-dotenv
sqlalchemy
alembic
geopy
tzwhere
pandas
seaborn
shapely
requests